requirements.txt
# Core framework dependencies
langchain==0.1.0
langchain-community==0.0.13
langgraph==0.0.22
langsmith==0.0.83

# Web automation
playwright==1.41.0

# Data validation and parsing
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM providers
openai==1.7.2
anthropic==0.8.1

# Database and storage
aiosqlite==0.19.0
asyncpg==0.29.0

# HTTP client
httpx==0.26.0

# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Authentication and security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Image processing for screenshots
Pillow==10.2.0

# Logging and monitoring
structlog==23.2.0

# Environment and configuration
python-dotenv==1.0.0

# Utilities
typing-extensions==4.9.0
uvloop==0.19.0

# Development dependencies
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0
black==23.12.1
isort==5.13.2
mypy==1.8.0
pre-commit==3.6.0